include: arch32: Fix the incorrect type
The temp val should be uint64_t in read_sysreg64 instead of uint32_t.
The incorrect type uint32_t will cause an issue:
mrrc instruction needs 2 registers when reading from a 64-bit system
register. The type uint32_t tells GCC only to save/restore one register,
so after the mrrc is executed, the other would clash.

+++ include/arch/arm/aarch32/cortex_a_r/lib_helpers.h
@@ -30,7 +30,7 @@
 
 #define read_sysreg64(op1, CRm)						\
 ({									\
-	uint32_t val;							\
+	uint64_t val;							\
 	__asm__ volatile ("mrrc p15, " #op1 ", %Q0, %R0, c"		\
 			  #CRm : "=r" (val) :: "memory");		\
 	val;								\
